factor out bottom navigation bar
